+++ lwk_wasm/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Async constructors for `Jade`, `JadeWebSocket`, and `Amp0Connected` where deprecated due to [wasm-bindgen/wasm-bindgen#4402](https://github.com/wasm-bindgen/wasm-bindgen/pull/4402)
+  * To avoid generating invalid TS code, macros `skip_typescript` was added to all of them.
+  * Created static async factories `Jade.fromSerial(...)`, `JadeWebSocket.fromWebSocket(...)`, and `Amp0Connected.connect(...)` to replace deprecated constructors
+
 ## 0.16.0
 
 * `TxOutSecrets` no longer exposes the asset blinding factor as a plain hex string in the old wasm API shape.